+++ src/app/routes/accounts.ts
@@ -53,6 +53,8 @@ accountsRouter.post(
         try {
             const account = await pecorino.service.account.open({
                 project: req.body.project,
+                // 互換性維持対応として、未指定であれば'Account'
+                typeOf: (typeof req.body.typeOf === 'string') ? req.body.typeOf : 'Account',
                 accountType: req.body.accountType,
                 accountNumber: req.body.accountNumber,
                 name: req.body.name,