+++ src/app/routes/transactions/withdraw.ts
@@ -67,10 +67,6 @@ withdrawTransactionsRouter.post(
             .isEmpty()
             .withMessage(() => 'required')
             .isInt(),
-        body('accountType')
-            .not()
-            .isEmpty()
-            .withMessage(() => 'required'),
         body('fromAccountNumber')
             .not()
             .isEmpty()
@@ -98,8 +94,6 @@ withdrawTransactionsRouter.post(
                     clientUser: req.user,
                     amount: parseInt(req.body.amount, 10),
                     fromLocation: {
-                        typeOf: pecorino.factory.account.TypeOf.Account,
-                        accountType: req.body.accountType,
                         accountNumber: req.body.fromAccountNumber
                     },
                     description: (req.body.notes !== undefined) ? req.body.notes : ''