+++ CHANGELOG.md
@@ -13,10 +13,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - update @pecorino/factory
 - 口座番号をグローバルユニークに変更
 - 口座に対する処理を口座番号の指定のみで実行するように調整
+- 処理は口座のtypeOfに依存しないように調整
 
 ### Deprecated
 
 ### Removed
+
+- ベーシック認証設定を削除
 
 ### Fixed